rename util folder
